Add google_sign_in feature and update login layout
