+++ .vscode/launch.json
@@ -1,0 +1,28 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "jam_scene_UI",
+            "cwd": "jam_scene_UI",
+            "request": "launch",
+            "type": "dart"
+        },
+        {
+            "name": "jam_scene_UI (profile mode)",
+            "cwd": "jam_scene_UI",
+            "request": "launch",
+            "type": "dart",
+            "flutterMode": "profile"
+        },
+        {
+            "name": "jam_scene_UI (release mode)",
+            "cwd": "jam_scene_UI",
+            "request": "launch",
+            "type": "dart",
+            "flutterMode": "release"
+        }
+    ]
+}
