+++ jam_scene_UI/ios/Flutter/Generated 3.xcconfig
@@ -1,0 +1,14 @@
+// This is a generated file; do not edit or check into version control.
+FLUTTER_ROOT=/Users/jnees/development/flutter
+FLUTTER_APPLICATION_PATH=/Users/jnees/Documents/OSU/467/capstone/jam_scene_UI
+COCOAPODS_PARALLEL_CODE_SIGN=true
+FLUTTER_TARGET=/Users/jnees/Documents/OSU/467/capstone/jam_scene_UI/lib/main.dart
+FLUTTER_BUILD_DIR=build
+FLUTTER_BUILD_NAME=1.0.0
+FLUTTER_BUILD_NUMBER=1
+EXCLUDED_ARCHS[sdk=iphonesimulator*]=i386
+DART_DEFINES=RkxVVFRFUl9XRUJfQVVUT19ERVRFQ1Q9dHJ1ZQ==
+DART_OBFUSCATION=false
+TRACK_WIDGET_CREATION=true
+TREE_SHAKE_ICONS=false
+PACKAGE_CONFIG=/Users/jnees/Documents/OSU/467/capstone/jam_scene_UI/.dart_tool/package_config.json
