+++ jam_scene_UI/ios/Flutter/flutter_export_environment 3.sh
@@ -1,0 +1,14 @@
+#!/bin/sh
+# This is a generated file; do not edit or check into version control.
+export "FLUTTER_ROOT=/Users/jnees/development/flutter"
+export "FLUTTER_APPLICATION_PATH=/Users/jnees/Documents/OSU/467/capstone/jam_scene_UI"
+export "COCOAPODS_PARALLEL_CODE_SIGN=true"
+export "FLUTTER_TARGET=/Users/jnees/Documents/OSU/467/capstone/jam_scene_UI/lib/main.dart"
+export "FLUTTER_BUILD_DIR=build"
+export "FLUTTER_BUILD_NAME=1.0.0"
+export "FLUTTER_BUILD_NUMBER=1"
+export "DART_DEFINES=RkxVVFRFUl9XRUJfQVVUT19ERVRFQ1Q9dHJ1ZQ=="
+export "DART_OBFUSCATION=false"
+export "TRACK_WIDGET_CREATION=true"
+export "TREE_SHAKE_ICONS=false"
+export "PACKAGE_CONFIG=/Users/jnees/Documents/OSU/467/capstone/jam_scene_UI/.dart_tool/package_config.json"
