+++ jam_scene_UI/lib/components/logo.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class Logo extends StatelessWidget {
+  const Logo({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Image.asset(
+      'assets/images/JamSceneLogoRaised.png',
+      height: MediaQuery.of(context).size.height * .35,
+    );
+  }
+}
