+++ jam_scene_UI/pubspec.yaml
@@ -40,6 +40,7 @@ dependencies:
   firebase_core_platform_interface: ^4.4.1
   flutter_svg: ^1.1.1+1
   google_sign_in: ^5.3.3
+  flutter_signin_button: ^2.0.0
 
 dev_dependencies:
   flutter_test:
